Improve installer.sh file

+++ install.sh
@@ -1,7 +1,46 @@
-brew install neovim/neovim/neovim
+  # The package is installed
+  brew install neovim
 
-ln -s ~/.config/nvim/bash_profile ~/.bash_profile
-ln -s ~/.config/nvim/tmux.conf ~/.tmux.conf
+read -p "Do you want to install .bash_profile config file?(y/n)" -n 1 -r
+echo
+
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+  if [ -f ~/.bash_profile ]; then
+    echo ".bash_profile already exist.. performing a backup before link the .bash_profile"
+    mv ~/.bash_profile ~/.bash_profile.old
+    echo "your previous configuration was renamed as ~/.bash_profile.old"
+  fi
+
+  ln -s ~/.config/nvim/bash_profile ~/.bash_profile
+  echo "bash_profile linked correctly!"
+fi
+
+read -p "Do you want to install .tmux.conf config file?(y/n)" -n 1 -r
+echo
+
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+  if [ -f ~/.tmux.conf ]; then
+    echo ".tmux.conf already exist.. performing a backup before link the .tmux.conf"
+    mv ~/.tmux.conf ~/.tmux.conf.old
+    echo "your previous configuration was renamed as ~/.tmux.conf.old"
+  fi
+
+  ln -s ~/.config/nvim/tmux.conf ~/.tmux.conf
+  echo "tmux.conf linked correctly!."
+fi
+
+read -p "Do you want to install .bash_aliases config file?(y/n)" -n 1 -r
+echo
+
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+  ln -s ~/.config/nvim/bash_aliases ~/.bash_aliases
+  echo ".bash_aliases linked correctly!."
+fi
+
+echo "Installing vim Plug plugin"
+
 curl -fLo ~/.config/nvim/autoload/plug.vim --create-dirs \
     https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
 nvim +BundleInstall +qall
+
